frontend: fixed the context not working
