+++ frontend/src/components/GameActionSelector.tsx
@@ -2,7 +2,7 @@ import { useMutation } from "@tanstack/react-query"
 import { AlertTriangle, CheckCircle2, Coins } from "lucide-react"
 import type { KeyboardEvent, ReactElement } from "react"
 import type * as ApiTypes from "@api-types"
-import type { PlayGameState } from "../routes/play.$gameId.tsx"
+import type { PlayGameState } from "../contexts/PlayGameContext.tsx"
 import { useBackendApiClient } from "../contexts/BackendApiClientContext.tsx"
 import { cn } from "../lib/cn.ts"
 import { Alert, AlertDescription, AlertTitle } from "./ui/alert.tsx"
