+++ frontend/src/components/GameLayout.tsx
@@ -1,6 +1,6 @@
 import type { ReactElement, ReactNode } from "react"
 import type * as ApiTypes from "@api-types"
-import type { PlayGameState } from "../routes/play.$gameId.tsx"
+import type { PlayGameState } from "../contexts/PlayGameContext.tsx"
 import { GameSideNav } from "./GameSideNav.tsx"
 import { GameTopBar } from "./GameTopBar.tsx"
 import { Skeleton } from "./ui/skeleton.tsx"
