+++ frontend/src/routes/play.$gameId.index.tsx
@@ -1,11 +1,11 @@
 import { createFileRoute } from "@tanstack/react-router"
 import type { ReactElement } from "react"
-import { StarMapPlaceholder } from "./play.$gameId.map.tsx"
+import { GameStarMap } from "../components/GameStarMap.tsx"
 
 export const Route = createFileRoute("/play/$gameId/")({
   component: PlayGameIndex,
 })
 
 function PlayGameIndex(): ReactElement {
-  return <StarMapPlaceholder />
+  return <GameStarMap />
 }
